requirements.txt
cloudify-plugins-common==3.2
cloudify-rest-client==3.2
cloudify-fabric-plugin==3.2
